python: Fix type checking when using the `@function` decorator
The fix was to cast the `function` decorator's return type to the wrapped function instead of the `Function` descriptor. Not fixed 100% though since it's not handling the use case of using another object's constructor in a method (e.g., `foo = function(Foo)`), but that's a very recent feature that's not released yet.

Renamed `TypeVar("F")` to `R` for clarity so that `Func[P, R]` reads like "function with parameters “P” and return type “R”".

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>