core: set endpoint map when initializing APIServer
Without this we're getting nil panics when using dagger shell.

Went unnoticed because we haven't added test coverage for dagger shell
to the integ tests yet due to the fact that testing tty programs in our
current testing setup is very non-obvious.

I will add whatever best test I can in a follow up to this PR, but
sending the fix separately since it is quite trivial and probably an
order of magnitude less time consuming than figuring out a test.

Signed-off-by: Erik Sipsma <erik@dagger.io>