go: decouple type parsing from code generation in module
This is a small-ish scale refactor that results in the code we use to
parse various types/functions/etc. in the go module codegen code.
There should be no functional changes, just code refactor+cleanup.

The main motivation is to support the interface work, where we need to
use parsed type information for more than just codegen. Before this
change I had just copy pasted hundreds of lines of code and slightly
tweaked them; this refactor felt like the best relatively quick approach
that wasn't unreasonable.

The secondary goal is just general cleanup of this code, which has been
growing organically+quickly as we've been iterating on it. This change
is far from a full cleanup, but does nudge it in the right direction by
detangling a lot of large blobs of code w/ mixed parsing and codegen and
splitting out to separate files a bit more.

Signed-off-by: Erik Sipsma <erik@dagger.io>