chore: simplify internal/mage imports
This package moves out the constants that are shared between the engine
and the build system (used for have common shared paths/etc) to be owned
by the top-level `github.com/dagger/dagger` package instead of the
internal `github.com/dagger/dagger/internal/mage` package.

This allows us to invert the dependency, so that the build system relies
on the engine, and not the other way round - having it the other way was
confusing, the build system should depend on the code, not the other way
round.

Because of this, we can remove the empty
v0.0.0-00010101000000-000000000000 version for `internal/mage` from our
main `go.mod` - we couldn't previously set this to be a sane version
number - our CI isn't versioned like that. This invalid version number
meant we couldn't import any package that eventually depended on
`internal/mage` from a completely external go package (more of a
weirdness, than an actual issue):

	$ go mod init main
	$ cat main.go
	package main

	import (
		"fmt"

		"github.com/dagger/dagger/core/schema"
	)

	func main() {
		fmt.Println(schema.Cache)
	}
	$ go mod tidy
	go: downloading github.com/dagger/dagger/internal/mage v0.0.0-00010101000000-000000000000
	go: main imports
		github.com/dagger/dagger/core/schema imports
		github.com/dagger/dagger/internal/mage/consts: github.com/dagger/dagger/internal/mage@v0.0.0-00010101000000-000000000000: invalid version: unknown revision 000000000000

With this patch, the above can actually be done, since the
`internal/mage` dependency is removed.

Signed-off-by: Justin Chadwell <me@jedevc.com>