Backfill integ tests for dagger shell + assorted fixes (#6251)
* core: fix goroutine leak in shell service

We weren't closing the side of the stdio pipes used by the shell
handler, which appeared to result in the goroutines used to copy data to
leak in some circumstances. Noticed while getting stack traces of the
engine during debugging of other issues.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* shell: support shell from nested sessions

Previously dagger shell did not work at all from nested execs or from
under "dagger run". This used to be an obscure case but now matters
since we want to add integ tests for shell.

The first problem was that the websocket http request made by the CLI
wasn't sending the HTTP basic auth headers w/ the session token.

The second problem was that ServeHTTP in the engine client only proxied
simple http requests, which doesn't cover the case where the conn gets
hijacked during a websocket upgrade. We now handle that case by
identifying websocket requests and proxying the underlying net.Conns
directly.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* tests: add coverage for dagger shell

Signed-off-by: Erik Sipsma <erik@dagger.io>

---------

Signed-off-by: Erik Sipsma <erik@dagger.io>