fix: omit buildkit/createdat timestamp from blob source
When calling GetRemotes to get a descriptor for the local source, so we
can put into our custom BlobSource, buildkit creates a custom
buildkit/createdat annotation on the descriptor that represents the
timestamp that the blob was created at.

By default, we were pulling all of the descriptor annotations into the
source identifier, which as a result changed the LLB definition (and the
vertex digest, which obviously causes all of the dependents to change as
well). This can be observed in the following conditions:

1. Run a build using a local directory, and observe the blob id
2. Restart the engine, and use a clean cache
3. Run another build using the same (unchanged) local directory, and
   observe a different blob id - since the directory is unchanged, the
   only difference here is in the buildkit/createdat annotation.

We don't actually need this annotation for anything: the descriptor is
only ever used in Snapshot, as an argument to GetByBlob. GetByBlob is
implemented on CacheManager in BuildKit, where only the digest,
mediaType, size and containerd uncompressed annotation are required.

As a result, we can avoid having the blob vertex digest continually
shift, by just filtering the annotations to only the ones we know we
need - at the moment, just containerd.io/uncompressed.

Signed-off-by: Justin Chadwell <me@jedevc.com>