feat: lay foundations for engine drivers
This patch starts to simplify the connhelpers and docker-image helper
into specific engine drivers.

Each driver registers itself under a name, and is called when connecting
to a URL with that scheme. Additionally, during connection, we provide a
set of additional parameters.

This is mostly just a refactor around the connhelpers, which we can't
use directly out of the box, since we want to support a set of
additional parameters as well. Since each driver just returns a
net.Conn, then we can build the buildkit connection on top of it, which
allow us to potentially change the connection mechanism in the future.

Signed-off-by: Justin Chadwell <me@jedevc.com>