tests: run dagger shell test in isolation + pre-cached (#6284)
* tests: run dagger shell test in isolation + pre-cached

Saw while reviewing PRs that the TestModuleDaggerShell test was flaking
sometimes specifically in the testdev workflow. Have never seen this
flake in the other "normal" integ test workflow or locally.

From the logs I saw it looks like it is running as expected but just
extremely slow, to the point that the generous timeouts weren't enough.

The testdev workflow is extra slow and overloaded for various unrelated
reasons (doesn't actually run on the shared runner and doesn't have the
same parallelism limits in the engine set).

Since I can't repro, this change is just an attempt at accounting for
those tests being really extra slow. It:
1. Causes the Shell test to not run in parallel with the others, so it
   should be less bogged down by other noise.
2. Calls "functions" on the module first so the module load is cached
   before we actually call "shell" (from the log output I saw, it seemed
   like that module load part was the one taking forever and causing the
   shell test timeout to get hit)

Signed-off-by: Erik Sipsma <erik@dagger.io>

* try reducing size of tty window

Signed-off-by: Erik Sipsma <erik@dagger.io>

---------

Signed-off-by: Erik Sipsma <erik@dagger.io>