core: add currentTypeDefs api for introspecting TypeDefs.
This enables callers to directly introspect a full loaded schema on the
level of TypeDefs rather than the introspection json. This includes the
core API, which is what the bulk of the new code here is supporting.

This is intended to be used by the CLI to support exposing the core API
the same way the way user module objects+functions currently are.

Signed-off-by: Erik Sipsma <erik@dagger.io>