fix: skip `head merge commits` on pull requests (#6253)
* fix: skip Head merge commits

When collecting labels on the engine in a VCS context, sometimes, the commit head is an appended commit (from that VCS)

Signed-off-by: grouville <guillaume@dagger.io>

* fix: Update TestLoadGitLabels to account for branch labels in detached HEAD state

The TestLoadGitLabels test was updated to account for the `dagger.io/git.branch` label in the expected output for the "detached HEAD state" scenario.

This change was necessary due to recent updates in the `LoadGitLabels` function in the `label.go` file. The function was modified to iterate over all references in the repository and check if any of them point to the commit hash. If a reference is found that matches the commit hash and it is a branch, the function adds the `dagger.io/git.branch` label.

This new behavior allows the function to correctly identify the branch association of a commit, even in a detached HEAD state. The test was updated to reflect this change and now expects the `dagger.io/git.branch` label in the output for the "detached HEAD state" scenario.

Signed-off-by: grouville <guillaume@dagger.io>

* fix: only skip merge commits on PRs

Following Justin's comment, only escape merge commits in PR context

Signed-off-by: grouville <guillaume@dagger.io>

* fix: do not skip merge requests on main branch + fix GitHub and GitLab

This commit implements a few changes:
- After further checks, GitLab does not change the git history in a merge request context. It doesn't need to be handled
- GitHub automatically creates a detached merge commit containing only the merge commit.
To handle it, we need to fetch the origin branch and retrieve this commit. However, this requires authentication, which we bypass by leveraging the fact that GitHub Action automatically adds the GITHUB_TOKEN to any git commands inside their runner context

Signed-off-by: grouville <guillaume@dagger.io>

* chore: burst dagger-runner cache

Signed-off-by: grouville <guillaume@dagger.io>

* chore: improve comment && burst cache

Signed-off-by: grouville <guillaume@dagger.io>

* feat: separate git fetch logic between origin remote and fork remote

Aim is to have a very readable implementation that handles the two separate logics:
- When the PR comes from a branch on the repo itself (origin)
- When the PR comes from a branch on a fork. Here, GitHub does not give us enough information to find with certainty the remote. So we try to construct the remote from the env var information. If it does not work, continue with current implementation

Signed-off-by: grouville <guillaume@dagger.io>

* feat: limit depth to 1 (HEAD) when fetching remote

In order to be as fast as possible, only fetch HEAD on either forked or origin remote

Signed-off-by: grouville <guillaume@dagger.io>

* feat: remove tmate debugging

Remove tmate for debugging purposes

Signed-off-by: grouville <guillaume@dagger.io>

* feat: handled self hosted / github entreprise repos

Signed-off-by: grouville <guillaume@dagger.io>

---------

Signed-off-by: grouville <guillaume@dagger.io>