chore: remove dockerfile
As far as I'm aware, this Dockerfile isn't actually used for anything?
It's not used in our own internal/mage package, and it isn't referenced
anywhere in our docs, I think this is legacy from before dagger was
built in dagger? Also, since go 1.21, this dockerfile no longer even
builds! (gotta love the best error message of all time `invalid go
version '1.21.5': must match format 1.23`)

Additionally, the constants for go versions and alpine are out-of-date
from the rest of the code.

Signed-off-by: Justin Chadwell <me@jedevc.com>