fix: avoid long holds on services mutex
Because Service.Stop has become a blocking function, we need to refactor
to ensure that we don't unneccessarily hold the lock.

This was preventing us from calling .Stop multiple times - e.g. if we
were to call with force=false, and then again with force=true, the
force=true would not be able to acquire the lock, and so the process
might not terminate.

Signed-off-by: Justin Chadwell <me@jedevc.com>