Groundwork for exposing core API in CLI (#6293)
* core: add currentTypeDefs api for introspecting TypeDefs.

This enables callers to directly introspect a full loaded schema on the
level of TypeDefs rather than the introspection json. This includes the
core API, which is what the bulk of the new code here is supporting.

This is intended to be used by the CLI to support exposing the core API
the same way the way user module objects+functions currently are.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* cli: add call experimental flag to enable inclusion of core types

We want this to eventually become the default behavior of a unified
dagger CLI command for invoking module functions, but in the mean time
this experimental flag allows us to iterate on the UX and features
needed without breaking or impacting existing functionality.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* cli: default call/download/up/shell to have core APIs

We can actually do this in the short term without breaking the current
UX. The modification needed was to change the OnSelectObjectLeaf
callback to run once the last command is being executed (i.e. we're
about to construct and submit the actual API query).

This allows the existing behavior to be retained (for now) while also
supporting continued chaining off of core object types like
Container/Directory/File too.

As part of this I also split the existing `dagger call` integ tests to
their own file, backfilled some coverage around the current behavior
when Container/Directory/File is the final return type and also added
new coverage for the new core chaining abilities.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* tests: backfill coverage for `dagger up`.

Wanted to have this to ensure the CLI changes don't break the up
command.

The test currently only uses `--native` because I hit a pre-existing bug
with `--port` (linked to the GH issue opened for it in the test case
code).

Signed-off-by: Erik Sipsma <erik@dagger.io>

* fix typedef return when scalar id

Signed-off-by: Erik Sipsma <erik@dagger.io>

* appease linter

Signed-off-by: Erik Sipsma <erik@dagger.io>

---------

Signed-off-by: Erik Sipsma <erik@dagger.io>