go sdk: comment the new codegen w/ docs+examples
Follow-up for previous PR that added a lot of this fairly intricate new
codegen.

I was a bit torn on whether to include as many concrete examples in the
comments since it technically adds a small maintainance burden to keep
the comments up to date whenever we change things.

However, in practice as working on this code I always ended up writing
temporary throw-away comments like this (to help reason through the code
to write) and also end up looking at the generated code that results. So
I think keeping the comments up to date is relatively straightforward
and worth the benefits they provide in terms of grounding the dense
jennifer code with something less "meta" to look at.

Signed-off-by: Erik Sipsma <erik@dagger.io>