Don't print the full env var name in error message.
This avoids a possible credential leak, in case the user accidentally passes the env variable instead of its name. This is particularly likely because until very recently, that's exactly what we told users to do. So there is a real risk of leaks for early adopters, which this change mitigates.

Signed-off-by: Solomon Hykes <solomon@dagger.io>