chore: ensure git clone step in sdk publish always runs
This step should *always* run - otherwise, it can accidentally cache,
and we wouldn't run the rest of the pipeline, even when we would have
new tags/branches/commits.

The "correct" solution long term is to ensure that we use the top-level
`client.Git` method here - however, due to a weird upstream bug, we end
up confusing tags and branches when cloning, which causes the
filter-branch command we use to fail. (I'm following up with this
upstream).

Signed-off-by: Justin Chadwell <me@jedevc.com>