Use a more recent version of CLI when installing via script
0.6.2 is really old (June 2023) and no longer supported, we don't want
to be giving it as an example.

Another tip is to use `BIN_DIR` instead of needing to `cd` before
installing.

Also, avoid using `/usr/local` for installing our binary. That is a
global path which requires special privileges. `$HOME/.local/bin` is a
much better default:
https://unix.stackexchange.com/questions/316765/which-distributions-have-home-local-bin-in-path

Lastly, drop `-L` flag for curl. We no longer need it since we switched to dl.dagger.io.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>