fix overzealous validation
In the previous commit I added a bunch of new tests and some extra
validation, but then forgot to run the rest of the tests, which were
failing with validation errors on paths in cases that were actually
okay.

This fixes the validation to allow relative paths provided they don't
escape the root dir, which requires some careful handling to deal with
the fact that e.g. filepath.Join(/, "../..") == "/".

Signed-off-by: Erik Sipsma <erik@dagger.io>