Support engine versioning by commit+local changes.
This fixes some bugs and improves how the engine is versioned in the
case of non-tagged builds (both off the main branch and local builds).

Now, when we build+publish the engine off of main the Version embedded
into the engine is set to the git commit it was built from. We also
publish the engine image with an image tag set to that commit rather
than "main". This is something we've wanted to do for a while and is
helpful for being able to easily use CLIs built off of main with the
right engine (avoiding the need for explicit _DAGGER_RUNNER_HOST
settings by the user).

Previously, there was also a bug in that we didn't actually include the
.git dir during builds and thus the vcs.revision we tried to read was
never actually available.

Now when a local dev engine is built the Version string is instead set
to a content hash of the state of the repo, including any
non-staged+non-committed modifications. This is helpful setup for the
next commits which help ensure that nested exec execution is always
invalidated when the engine changes.

Signed-off-by: Erik Sipsma <erik@dagger.io>