Invalidate nested exec cache if engine version changes.
Before this, any change to the implementation of the engine did not
result in nested execs (which can call back to the engine API) being
invalidated.

Dagger developers working on the engine locally hit this all the time
since you frequently needed to explicitly clear the dev engine's cache
in order for Module tests to actually run with the latest changes made
locally.

This same problem could also generally impact any user though as an
upgrade to a new engine version could still result in module execution
from other engines being re-used in cache.

Now, the embdedded engine.Version string is in the cache key of nested
execs.

This fixes the problem for official tagged releases, builds off of main
and local builds with unstaged modifications due to the previous commits
that fixed engine.Version to be set based on all those factors.

Signed-off-by: Erik Sipsma <erik@dagger.io>