fix: avoid holding server map lock
We should avoid holding the server map lock for any long period of time.
This should help to prevent a total deadlock in the case where a rogue
server holds the lock for too long during shutdown (for currently
unknown reasons).

By reducing our usage of the global map lock to only simple operations
that won't block, we ensure that this rogue server doesn't block every
operation on the controller.

However, because of this refactor, we need to introduce a more precise
per-server-id lock. This lock is used to prevent multiple incoming
session requests from spawning multiple servers for the same ID in
parallel.

Signed-off-by: Justin Chadwell <me@jedevc.com>