cache: make first synchronous import non-fatal and timed out.
Before this, if there was an error when synchronously importing layer
cache at engine start, the engine failed to start entirely. Now, the
engine will continue running with just the local cache and retry import
periodically in the background.

Additionally, that initial synchronous import did not have a timeout
previously. It now has a timeout of 1 minute. If that's hit, then the
above logic kicks in to gracefully handle the error.

Signed-off-by: Erik Sipsma <erik@dagger.io>