chore: create test client at lowest Parallel level
We should avoid calling Parallel in a sub-test *after* creating a
client, since it can block for a completely indeterminate amount of
time, and in that time, we can end up with a huge number of just unused
clients.

Not only is this (really only a) little wasteful, it makes stack traces
and dumps in the case of test timeout very hard to decipher. Are we
waiting for a client because it got deadlocked? Or because it was open
since the beginning of the test run.

Signed-off-by: Justin Chadwell <me@jedevc.com>