Update zenith dev guides with incoming breaking changes.
The most important updates here are in anticipation of breaking changes
in the next release. There are also some more opinionated cleanups, but
for now just leaving the overall structure the same.

The big breaking changes to adapt for are:
1. `dagger mod init/install/etc.` moving to be right under `dagger`
2. `dagger up` replaced by chainable `dagger call ... up`
3. `dag.Host()` being replaced with `dag.CurrentModule().Source()`

The more opinionated tweaks are:
1. Don't use raw graphql queries, only use `dagger call`
2. In Go, avoid use of `Optional` and instead center on `//+optional`
   and `//+default` pragmas.

One important note: the moving of `dagger mod init/install/etc.` has not
been merged yet, so depending on how the timing works out may or may not
need to tweak this update before merging.

Signed-off-by: Erik Sipsma <erik@dagger.io>