Revert "rip out progrock socket plumbing"
Turns out this is still needed by legacy dagger (dagger run, go run +
auto-started session, etc).

It's unfortunate to bring all this plumbing back but we obviously can't
break all that yet.

Kept everything as-is, including propagating parents, because #6505
prevents the cache-busting issue that we originally ripped all this out
to avoid.

This reverts commit 3309a8e799e1f4d66506bdc5e540613ae4bab90a.

Signed-off-by: Alex Suraci <alex@dagger.io>