The new dag.CurrentModule() API replaces the need for the Host API and gives more controlled functionality that is specific to modules.
We do need to leave in the underlying graphql schema for host because
it's needed for internal calls (e.g. a host service passed to a module
function). So this just results in the Host types being scrubbed from
the introspection JSON that is provided to module SDKs for codegen.

Signed-off-by: Erik Sipsma <erik@dagger.io>