core: fix support for custom SDK modules (#6545)
The recent updates accidentally left in dag.Select calls that referred
to old APIs in the codepath for custom (i.e. 3rd party) SDKs. We didn't
catch this because we didn't have any integ tests for this feature.

This fixes the bug and adds an integ test. It also required exposing
Module.Runtime as a graphql field in order to write the test case, but
that seems like it will only be beneficial in general to any other SDK
authors trying to do something similar to this.

Signed-off-by: Erik Sipsma <erik@dagger.io>