Remove Host API from module SDK codegen. (#6535)
* The new dag.CurrentModule() API replaces the need for the Host API and
gives more controlled functionality that is specific to modules.

We do need to leave in the underlying graphql schema for host because
it's needed for internal calls (e.g. a host service passed to a module
function). So this just results in the Host types being scrubbed from
the introspection JSON that is provided to module SDKs for codegen.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* tweak introspection serialization for python graphql-core

This library seems to be very opinionated about the exact format of the
introspection json, so need to add a few `omitempty`s and add a
(currently always empty) `interfaces` to `Type`.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* add changie log w/ breaking change notice

Signed-off-by: Erik Sipsma <erik@dagger.io>

* chore: remove changelog entry for modules

Signed-off-by: Justin Chadwell <me@jedevc.com>

---------

Signed-off-by: Erik Sipsma <erik@dagger.io>
Signed-off-by: Justin Chadwell <me@jedevc.com>
Co-authored-by: Justin Chadwell <me@jedevc.com>