chore: refactor customer dns sources to avoid hacks
This removes our smuggling hack where we smuggle properties through
existing ones - we no longer need to do this, since we can create a
custom identifier that wraps the upstream ones.

This is hopefully part 1 of several to try and clean up our custom
sources - I'd love to be able to remove most of the copy-pasta code from
buildkit, to avoid future drift.

Signed-off-by: Justin Chadwell <me@jedevc.com>