fix: gracefully handle StartBindings failure
While investigating something (hopefully) unrelated to this, I found
that StartBindings wasn't handling error cases well. There is an attempt
to try and call `detach` on failure, but it wasn't being used.

Because we use a channel to pull the binding results, if one of the
services failed to start for some reason,  we would exit early, and none
of the started services would be in `running`, so when we called
`detach`, none of the services would be stopped. This could easily
result in dangling services.

Signed-off-by: Justin Chadwell <me@jedevc.com>