add integ test for mod init --name and --sdk required together
Coverage for this was missing. I expected going into this that it would
actually not work as expected yet based on user reports but I haven't
been able to repro the behavior they saw yet. Will update if any bugs
are found. But in the meantime obviously need to have coverage of this
either way.

Signed-off-by: Erik Sipsma <erik@dagger.io>