update tests picked up from rebase to use dagger init
This also required a slight implementation adjustment to give
re-configuring an SDK the same behavior as re-configuring a name in that
it only becomes an error once you try to re-generate the source code.
This allows custom SDKs to re-use other SDKs more easily, which is
relied on for our custom SDK test.

Signed-off-by: Erik Sipsma <erik@dagger.io>