better step display heuristics
goals:

* avoid flickering by only removing steps after 1 second
* only show steps that took >100ms to avoid an overwhelming amount of
  noise

justification:

This new TUI should converge to only showing things that are actively
running. For example, 'dagger up' output should be nice and clean, only
showing the tunnels that were created and the services that are running.

We don't want the user to have to hand-configure hints for what to show,
and we don't want to accidentally hide critical things like "stuck"
steps. So we'll treat the duration of the step itself as the heuristic.

In the case of `dagger call`, the value you're querying will already be
printed on exit. Any other output is extra; the only reason you would
care about it is when it's taking too long. So, we only show what's
taking long.

Signed-off-by: Alex Suraci <alex@dagger.io>