avoid stack overflow while processing steps
this one's weird. it's not an obvious infinite recursion, but it seems
to be high enough to trip Go's stack limit. the problem goes away when
adding a runtime.Gosched() or log.Println in HighLevelStep.

turns out we don't even need to recurse, so now we just store a step's
parent digest on it, instead of its parent step.

Signed-off-by: Alex Suraci <alex@dagger.io>