core: fix native always true for up even when ports are set
The logic previously was accidentally setting native to true even when
ports were manually specified which resulted in each port being applied
twice. In the case where the same values of frontend and backend were
specified (i.e. 5000:5000) this resulted in an error since you'd try to
listen on the same port twice.

Signed-off-by: Erik Sipsma <erik@dagger.io>