Module engine version tracking (#6469)
* Support engine versioning by commit+local changes.

This fixes some bugs and improves how the engine is versioned in the
case of non-tagged builds (both off the main branch and local builds).

Now, when we build+publish the engine off of main the Version embedded
into the engine is set to the git commit it was built from. We also
publish the engine image with an image tag set to that commit rather
than "main". This is something we've wanted to do for a while and is
helpful for being able to easily use CLIs built off of main with the
right engine (avoiding the need for explicit _DAGGER_RUNNER_HOST
settings by the user).

Previously, there was also a bug in that we didn't actually include the
.git dir during builds and thus the vcs.revision we tried to read was
never actually available.

Now when a local dev engine is built the Version string is instead set
to a content hash of the state of the repo, including any
non-staged+non-committed modifications. This is helpful setup for the
next commits which help ensure that nested exec execution is always
invalidated when the engine changes.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* Add engine version to ModuleConfig in dagger.json.

Tracking engine version a module was developed/published with is useful
metadata to have in general, but it also in particular gives us an
option in the future to actually handle backwards incompatible changes
gracefully if we ever are compelled to.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* Invalidate nested exec cache if engine version changes.

Before this, any change to the implementation of the engine did not
result in nested execs (which can call back to the engine API) being
invalidated.

Dagger developers working on the engine locally hit this all the time
since you frequently needed to explicitly clear the dev engine's cache
in order for Module tests to actually run with the latest changes made
locally.

This same problem could also generally impact any user though as an
upgrade to a new engine version could still result in module execution
from other engines being re-used in cache.

Now, the embdedded engine.Version string is in the cache key of nested
execs.

This fixes the problem for official tagged releases, builds off of main
and local builds with unstaged modifications due to the previous commits
that fixed engine.Version to be set based on all those factors.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* wip fix tests after rebase

Signed-off-by: Erik Sipsma <erik@dagger.io>

* fix local imports to actually re-use refs

Signed-off-by: Erik Sipsma <erik@dagger.io>

* regen docs

Signed-off-by: Erik Sipsma <erik@dagger.io>

* remove Module.WithEngineVersion, just set value implicitly

Signed-off-by: Erik Sipsma <erik@dagger.io>

* shim: fix slice out of bounds panic in trie

Trie.Direct could be of len 0 but non-nil, so this just updates a check
to handle that case.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* fix typo in publish.yaml

Signed-off-by: Erik Sipsma <erik@dagger.io>

---------

Signed-off-by: Erik Sipsma <erik@dagger.io>