fix resolution of local dep from git module
The source root subpath of git refs were being turned into absolute
paths (as opposed to local refs which always stored the source root
subpath as relative). This broke the resolve dependency logic and
resulted in errors when installing modules from a git ref that had a
relative local dep.

Signed-off-by: Erik Sipsma <erik@dagger.io>