ci: Send Discord notifications when there is a new release
This adds notifications for the following releases:
1. 🚙 Engine + 🚗 CLI
2. 🐹 Go SDK
3. 🐍 Python SDK
4. ⬢ TypeScript SDK
5. ⚙️ Rust SDK (this is not published part of the regular release cycle)
6. 🧪 Elixir SDK
7. 🐘 PHP SDK
8. 🎡 Helm Chart (this is not published part of the regular release cycle)

We used to have a bot for this purpose: `github-release-towncrier`.
This required us to:
- keep the app running all the time
- change the filtering when a new release type was added (this explains
  why some release notifications are nto published today)

A few other issues that we've hit with this setup:
- performing any updates during a release would miss webhooks
- if a webhook was missed, the notification would not go out
- setting up redundant instances would result in multiple notifications
  for the same webhook

This new approach leverages Dagger Modules and doesn't require any code
to run: we simply invoke a remote module just-in-time.

https://github.com/gerhard/daggerverse/tree/2024-02-13/notify Dagger
Module was developed part of an internal Dagger Hackathon 🐶

See the PR for more details.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>