fix onlcr being "permanent" for terminal sessions
This took a long time and a ton of sleuthing to figure out.

Buildkit allocates a PTY for forwarding i/o over the frontend gateway.
Newly allocated PTS default to onlcr which means "translate \n to \r\n."
This is necessary for simple CLIs like bash and ls to write text
properly without it just walking across the screen.

TUIs like vim rely on terminfo to implement cursor positioning, and many
terminfo entries have cud1 ("move cursor down 1 cell") mapped to \n.
This is normally fine because TUIs like vim will usually put the PTY in
raw mode, which disables onlcr.

However, because we're running in a container, disabling onlcr was not
taking effect. So any time vim tried to move the cursor down one row
(\n), it would actually end up moving it to column 0 (\r) and then down
(\n).

The fix is to disable onlcr on the host side and then re-enable it on
the container side. That way we still have the sane default, and now we
can let programs disable it like they normally can.

Signed-off-by: Alex Suraci <alex@dagger.io>