fix: add generated go modules to go.work
This is a pretty common use case - we have a `go.work` at the root of a
repo, and we want to create a dagger module for it with `dagger init`.

However, naively, this will actually create a dagger module that can't
be called (because `go.work` doesn't include it). To resolve this, we
need to add it to `go.work`.

Signed-off-by: Justin Chadwell <me@jedevc.com>