terminal: preserve scrollback
this is a little tricksy! we use a new hook that Midterm calls with
lines as they go offscreen and queue them up for Bubbletea to print
above the view. that way we still have a fully dynamic renderable
viewport without stealing mouse input and without having to maintain an
infinitely growing terminal.

Signed-off-by: Alex Suraci <alex@dagger.io>