bump midterm for better TUI support (#6650)
* use strings.Builder, since we're building a string

Signed-off-by: Alex Suraci <alex@dagger.io>

* .

Seems like the convention is proper capitalization and punctuation. I
can vibe with that.

Signed-off-by: Alex Suraci <alex@dagger.io>

* bump midterm, more commentary

Signed-off-by: Alex Suraci <alex@dagger.io>

* fix terminal tests to not expect \n after prompts

Signed-off-by: Alex Suraci <alex@dagger.io>

* fix onlcr being "permanent" for terminal sessions

This took a long time and a ton of sleuthing to figure out.

Buildkit allocates a PTY for forwarding i/o over the frontend gateway.
Newly allocated PTS default to onlcr which means "translate \n to \r\n."
This is necessary for simple CLIs like bash and ls to write text
properly without it just walking across the screen.

TUIs like vim rely on terminfo to implement cursor positioning, and many
terminfo entries have cud1 ("move cursor down 1 cell") mapped to \n.
This is normally fine because TUIs like vim will usually put the PTY in
raw mode, which disables onlcr.

However, because we're running in a container, disabling onlcr was not
taking effect. So any time vim tried to move the cursor down one row
(\n), it would actually end up moving it to column 0 (\r) and then down
(\n).

The fix is to disable onlcr on the host side and then re-enable it on
the container side. That way we still have the sane default, and now we
can let programs disable it like they normally can.

Signed-off-by: Alex Suraci <alex@dagger.io>

* terminal: preserve scrollback

this is a little tricksy! we use a new hook that Midterm calls with
lines as they go offscreen and queue them up for Bubbletea to print
above the view. that way we still have a fully dynamic renderable
viewport without stealing mouse input and without having to maintain an
infinitely growing terminal.

Signed-off-by: Alex Suraci <alex@dagger.io>

* bump midterm to fix \e[?1006;1000h handling

Signed-off-by: Alex Suraci <alex@dagger.io>

---------

Signed-off-by: Alex Suraci <alex@dagger.io>