mark only actually internal ID vertex calls
previously marked all ID vertices as internal so they wouldn't be
displayed in non-Zenith UIs like dagger run or in Cloud.

now we will consider any `dag.Select` call as internal, and not set it
for IDs corresponding to direct user queries.

to maintain compatibility, the telemetry.LegacyIDInternalizer
progrock.Writer is used for Cloud and the old TUIs.

Signed-off-by: Alex Suraci <alex@dagger.io>