fix: always compute the right FromID name
We were previously incorrectly computing the wrong FromID name - this
was because we weren't taking into account the current namespace. We now
take this into account, by making sure that we always get the right
namespace.

The reason this never surfaced before was because we were manually
setting the `id` field, which meant we always returned it directly, and
never actually used the query object. This is changed now though, since
we can only set `id` on local interface implementations, and not foreign
structs (since those now live in a separate package, and there's not
really a *great* case for directly exposing `id`).

Signed-off-by: Justin Chadwell <me@jedevc.com>